requirements.txt
pydantic>=2.10.0
pillow>=11.2.1
requests>=2.32.0
tqdm>=4.67.0
rich>=13.9.0
qdrant-client>=1.12.0
pyyaml>=6.0.2
numpy>=2.2.0
typing-extensions>=4.12.2
pydantic-settings>=2.7.0
